pyproject.toml
[tool.poetry]
name = "tfs_viewer"
version = "0.1.0"
description = ""
authors = ["Felix Soubelet <felix.soubelet@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
streamlit = "^0.82.0"
pandas-profiling = "^3.0.0"
streamlit-pandas-profiling = "^0.1.2"
tfs-pandas = "^2.0.2"
watchdog = "^2.1.2"
plotly = "^4.14.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
